+++ src/Styles/addFd.css
@@ -1,0 +1,10 @@
+#addFd .deviceTable {
+    position: relative;
+}
+
+#addFd .processTable {
+    display: block;
+    width: 100%;
+    background-color: gray;
+    padding: 10px;
+}
