--- src/Styles/fileList.css
@@ -1,6 +1,0 @@
-.fileList {
-    width: 100%;
-    background-color: gray;
-    border-radius: 10px;
-    padding: 5px 5px 5px 5px;
-}
