--- src/Styles/processTable.css
@@ -1,6 +1,0 @@
-.processTable {
-    width: 100%;
-    background-color: #595959;
-    border-radius: 10px;
-    padding: 5px 5px 5px 5px;
-}
